add mutex lock to database state manager and move code to new file
